requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.1.7
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.3
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.3
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qrcode==8.1
redis==5.2.1
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
validate-docbr==1.10.0
vine==5.1.0
wcwidth==0.2.13
